chore: break out fetching logic to a specfic hook
